Update denormalized citizenparticipantcount when participating, creating or accepting invitation

+++ municipalityinitiative-web/src/main/resources/db/migration/V14__participant_count_citizen.sql
@@ -1,5 +1,5 @@
 ALTER TABLE municipality_initiative
-  ADD COLUMN participant_count_citizen BIGINT NOT NULL DEFAULT 0;
+  ADD COLUMN participant_count_citizen INTEGER NOT NULL DEFAULT 0;
 
 -- Denormalize from verified participants
 update municipality_initiative
